fix support for more that 4G tasks

+++ src/datawizard/data_request.h
@@ -1,6 +1,6 @@
 /* StarPU --- Runtime system for heterogeneous multicore architectures.
  *
- * Copyright (C) 2009-2010, 2013-2015  Université de Bordeaux
+ * Copyright (C) 2009-2010, 2013-2016  Université de Bordeaux
  * Copyright (C) 2010, 2011, 2013  CNRS
  *
  * StarPU is free software; you can redistribute it and/or modify
@@ -93,7 +93,7 @@ LIST_TYPE(_starpu_data_request,
 
 	struct _starpu_callback_list *callbacks;
 
-	unsigned com_id;
+	unsigned long com_id;
 )
 PRIO_LIST_TYPE(_starpu_data_request, prio)
